64 to 32

+++ models/Seq2seq.py
@@ -63,7 +63,7 @@ class Seq2seq(nn.Module):
 		noise = data_helpers.add_noise(sess, self.model, grad_noise,
                     src_ids, tgt_ids, embedding_dim, random_type=noise_config['noise_type'], 
                     word_keep=noise_config['word_keep'], weight=noise_config['weight'], mean=noise_config['mean'],
-					replace_map=noise_config['replace_map'])
+					replace_map=noise_config['replace_map']).astype(np.float32)
 		noise = torch.tensor(noise).to(device=device)
 		new_embeds = inputs_embeds * noise
 		pdb.set_trace()
